DEVPROPID is ULONG which maps to C#'s uint https://docs.microsoft.com/en-us/windows/win32/winprog/windows-data-types#ulong

+++ src/HidLibrary/NativeMethods.cs
@@ -118,7 +118,7 @@ namespace HidLibrary
         internal struct DEVPROPKEY
         {
             public Guid fmtid;
-            public ulong pid;
+            public uint pid;
         }
 
         internal static DEVPROPKEY DEVPKEY_Device_BusReportedDeviceDesc =
